+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/ParentSyncMessage.cs
@@ -6,7 +6,7 @@ namespace Unity.Netcode
     {
         public int Version => 0;
 
-        private const string k_Name = "DestroyObjectMessage";
+        private const string k_Name = "ParentSyncMessage";
 
         public ulong NetworkObjectId;
 
@@ -87,12 +87,20 @@ namespace Unity.Netcode
             reader.ReadValueSafe(out Rotation);
             reader.ReadValueSafe(out Scale);
 
-            // If the target NetworkObject does not exist =or= the target latest parent does not exist then defer the message
-            if (!networkManager.SpawnManager.SpawnedObjects.ContainsKey(NetworkObjectId) || (LatestParent.HasValue && !networkManager.SpawnManager.SpawnedObjects.ContainsKey(LatestParent.Value)))
+            // If the target NetworkObject does not exist then defer this message until it does.
+            if (!networkManager.SpawnManager.SpawnedObjects.ContainsKey(NetworkObjectId))
             {
                 networkManager.DeferredMessageManager.DeferMessage(IDeferredNetworkMessageManager.TriggerType.OnSpawn, NetworkObjectId, reader, ref context, k_Name);
                 return false;
             }
+
+            // If the target parent does not exist, then defer this message until it does.
+            if (LatestParent.HasValue && !networkManager.SpawnManager.SpawnedObjects.ContainsKey(LatestParent.Value))
+            {
+                networkManager.DeferredMessageManager.DeferMessage(IDeferredNetworkMessageManager.TriggerType.OnSpawn, LatestParent.Value, reader, ref context, k_Name);
+                return false;
+            }
+
             return true;
         }